requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
clear_screen==0.1.14
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
dotenv==0.9.9
executing==2.2.0
filelock==3.13.1
fonttools==4.59.0
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
noisereduce==3.0.3
numba==0.61.2
numpy==2.1.2
ollama==0.5.1
omegaconf==2.3.0
openai-whisper==20250625
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scipy==1.16.0
setuptools==70.2.0
simple-term-menu==1.6.6
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.3
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.5.0
wcwidth==0.2.13
